+++ lib/stream_cloudflare.dart
@@ -1,20 +1,22 @@
 import 'package:flutter/services.dart';
+import 'stream_cloudflare_platform_interface.dart';
 
 class StreamCloudflare {
   static const MethodChannel _channel = MethodChannel('stream_cloudflare');
 
-  /// Play video using Cloudflare Stream
   Future<void> play(String videoId) async {
     await _channel.invokeMethod('play', {'videoId': videoId});
   }
 
-  /// Set audio track
   Future<void> setAudioTrack(int trackId) async {
     await _channel.invokeMethod('setAudioTrack', {'trackId': trackId});
   }
 
-  /// Set subtitle track
   Future<void> setSubtitleTrack(int trackId) async {
     await _channel.invokeMethod('setSubtitleTrack', {'trackId': trackId});
+  }
+
+  Future<String?> getPlatformVersion() {
+    return StreamCloudflarePlatform.instance.getPlatformVersion();
   }
 }