+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- Initial release.
+- Support for playing videos from Cloudflare Stream.
+- Features include adaptive quality management, audio track switching, and subtitle support.